Add debug logs for nask * Delete funcs/nasksub.cpp * Referctoring for some codes
