+++ 20GO/libgo/fputs.cpp
@@ -7,7 +7,6 @@
 int GO_fputs(const char *s, GO_FILE *stream)
 {
 	unsigned int l = strlen(s);
-	#warning "Fix me"
-	//fwrite(s, 1, l, stream);
+	GO_fwrite(s, 1, l, stream);
 	return l;
 }
