+++ 20GO/toolstdc/Makefile.am
@@ -13,7 +13,7 @@ sjisconv_SOURCES= sjisconv.cpp
 gas2nask_SOURCES= gas2nask.cpp
 
 nask_SOURCES	= nask.cpp
-nask_CXXFLAGS   = -I$(top_builddir)/20GO/include
+nask_CXXFLAGS   = -I$(top_builddir)/20GO/include -I$(top_builddir)/20GO/libnask
 nask_LDADD	= $(top_builddir)/20GO/libnask/libnask.la $(top_builddir)/20GO/libgo/libgo.la
 
 sjisconv_CXXFLAGS = -I$(top_builddir)/20GO/include -I$(top_builddir)/20GO/libnask
