+++ 20GO/toolstdc/nask.cpp
@@ -1,7 +1,11 @@
 /* copyright(C) 2003 H.Kawai (under KL-01). */
 
-#include <stdio.h>
-#include <stdlib.h>
+#include <cstdio>
+#include <cstdlib>
+#include <iostream>
+#include "ll.hpp"
+#include "go_string.hpp"
+#include "../drv_stdc/others.hpp"
 
 #define SIZ_STDOUT			(16 * 1024)
 #define SIZ_STDERR			(16 * 1024)
@@ -26,10 +30,17 @@ extern struct GOL_STR_MEMMAN GOL_memman, GOL_sysman;
 int GOL_retcode;
 
 UCHAR *GOL_work0;
-int main1(int argc, UCHAR **argv, UCHAR *src0);
-void GOL_sysabort(unsigned char termcode);
-void *GOL_memmaninit(struct GOL_STR_MEMMAN *man, size_t size, void *p);
+UCHAR *LL(UCHAR *src0, UCHAR *src1, UCHAR *dest0, UCHAR *dest1);
+UCHAR *nask(UCHAR *src0, UCHAR *src1, UCHAR *dest0, UCHAR *dest1);
+UCHAR *output(UCHAR *src0, UCHAR *src1, UCHAR *dest0, UCHAR *dest1, UCHAR *list0, UCHAR *list1);
+int before_nask_process(int argc, UCHAR **argv, UCHAR *src0);
 int nask_main(int argc, UCHAR **argv);
+void *GOL_memmaninit(struct GOL_STR_MEMMAN *man, size_t size, void *p);
+void GOL_sysabort(unsigned char termcode);
+void errmsgout(const UCHAR* s);
+void errmsgout(const char* s);
+void errmsgout_s_NL(const UCHAR *msg, const UCHAR *s);
+void errmsgout_s_NL(const char* msg, const char* s);
 
 struct bss_alloc {
 	UCHAR _stdout[SIZ_STDOUT];
@@ -38,33 +49,177 @@ struct bss_alloc {
 	UCHAR work1[MAX_SRCSIZ + MAX_TMPSIZ + MAX_BINSIZ + MAX_LSTSIZ];
 };
 
-#include "../drv_stdc/others.hpp"
-
-// ƒƒCƒ“ŠÖ”‚Ì’è‹`‚Íã‘‚«‚Å‚«‚È‚¢
+// ãƒ¡ã‚¤ãƒ³é–¢æ•°ã®å®šç¾©ã¯ä¸Šæ›¸ãã§ããªã„
 int main(int argc, char** argv)
 {
+	LOG_DEBUG("nask argc: %d, argv:%s\n", argc, dump_argv(argv).c_str());
 	nask_main(argc, reinterpret_cast<UCHAR**>(argv));
 }
 
 int nask_main(int argc, UCHAR **argv)
 {
-	struct bss_alloc *bss0 = (struct bss_alloc *) malloc(sizeof (struct bss_alloc));
+	std::unique_ptr<struct bss_alloc> bss0(new bss_alloc());
 	GO_stdout.p0 = GO_stdout.p = bss0->_stdout;
 	GO_stdout.p1 = GO_stdout.p0 + SIZ_STDOUT;
 	GO_stdout.dummy = ~0;
 	GO_stderr.p0 = GO_stderr.p = bss0->_stderr;
-	GO_stderr.p1 = GO_stderr.p0 + (SIZ_STDERR - 128); /* ‚í‚´‚Æ­‚µ¬‚³‚­‚µ‚Ä‚¨‚­ */
+	GO_stderr.p1 = GO_stderr.p0 + (SIZ_STDERR - 128); /* ã‚ã–ã¨å°‘ã—å°ã•ãã—ã¦ãŠã */
 	GO_stderr.dummy = ~0;
 	GOL_memmaninit(&GOL_memman, SIZ_WORK, GOL_work0 = bss0->work);
 
-	GOL_retcode = main1(argc, argv, bss0->work1);
-	/* ƒoƒbƒtƒ@‚ğo—Í */
+	GOL_retcode = before_nask_process(argc, argv, bss0->work1);
+	LOG_DEBUG("nask GOL_retcode: %d", GOL_retcode);
+	/* ãƒãƒƒãƒ•ã‚¡ã‚’å‡ºåŠ› */
 	GOL_sysabort(0);
-	return 0; /* ƒ_ƒ~[ */
+	return 0; /* ãƒ€ãƒŸãƒ¼ */
 }
 
-// ‚È‚ñ‚Å‚±‚¤‚¢‚¤‚±‚Æ‚·‚é‚©‚È‚ HƒR[ƒh‚ğÄ—˜—p‚·‚éê‡Aƒ‰ƒCƒuƒ‰ƒŠ‚ğì‚é‚×‚«‚Å‚·‚æH
+// ãªã‚“ã§ã“ã†ã„ã†ã“ã¨ã™ã‚‹ã‹ãªã‚ï¼Ÿã‚³ãƒ¼ãƒ‰ã‚’å†åˆ©ç”¨ã™ã‚‹å ´åˆã€ãƒ©ã‚¤ãƒ–ãƒ©ãƒªã‚’ä½œã‚‹ã¹ãã§ã™ã‚ˆï¼Ÿ
 #include "../drv_stdc/others.cpp"
 #include "../drv_stdc/wfile_b.cpp"
 #include "../drv_stdc/wfile_t.cpp"
-#include "../funcs/nasksub.cpp"
+
+void errmsgout_s_NL(const char* msg, const UCHAR* s)
+{
+	return errmsgout_s_NL(reinterpret_cast<const UCHAR*>(msg), s);
+}
+
+void errmsgout(const char* s)
+{
+	return errmsgout(reinterpret_cast<const UCHAR*>(s));
+}
+
+void errmsgout(const UCHAR *s)
+{
+	int l =	GO_strlen(s);
+	char flag = 0;
+	GO_FILE *stream = &GO_stderr;
+	if (l >= stream->p1 - stream->p) {
+		l = stream->p1 - stream->p;
+		flag++;
+	}
+	if (l > 0) {
+		std::cerr << s << std::endl;
+	}
+	if (flag) {
+		GOL_sysabort(3 /* GO_TERM_ERROVER */);
+	}
+	return;
+}
+
+void errmsgout_s_NL(const UCHAR *msg, const UCHAR *s)
+{
+	std::cerr << msg << s << std::endl;
+	return;
+}
+
+int nask_errors = 0;
+
+int before_nask_process(int argc, UCHAR **argv, UCHAR *src0)
+{
+	UCHAR *src1, *dest0, *dest1;
+	UCHAR *tmp0, *tmp1, *list0, *list1;
+	int len;
+
+	tmp0 = src0 + MAX_SRCSIZ;
+	dest0 = tmp0 + MAX_TMPSIZ;
+	list0 = dest0 + MAX_BINSIZ;
+	if (argc < 2 || argc > 4) {
+		errmsgout("usage : >nask source [object/binary] [list]");
+		return 16;
+	}
+	len = GOLD_getsize(argv[1]);
+	LOG_DEBUG("GOLD_getsize(file): %d\n", len);
+	if (len == -1) {
+		errmsgout_s_NL("NASK : can't open ", argv[1]);
+		return 17;
+	}
+	if (len > MAX_SRCSIZ) {
+		errmsgout_s_NL("NASK : too large ", argv[1]);
+		return 17;
+	}
+	if (GOLD_read(argv[1], len, src0)) {
+		errmsgout_s_NL("NASK : can't read ", argv[1]);
+		return 17;
+	}
+	src1 = src0 + len;
+
+	LOG_DEBUG("call nask...\n");
+	list1 = nask(src0, src1, list0, list0 + MAX_LSTSIZ);
+	LOG_DEBUG("nask result: %s\n", list1);
+
+	if (list1 == NULL) {
+over_listbuf:
+		errmsgout("NASK : LSTBUF is not enough" NL);
+		return 19;
+	}
+
+	tmp1 = LL(list0, list1, tmp0, tmp0 + MAX_TMPSIZ);
+	if (tmp1 == NULL) {
+over_tmpbuf:
+		errmsgout("NASK : TMPBUF is not enough" NL);
+		return 19;
+	}
+
+	dest1 = output(tmp0, tmp1, dest0, dest0 + MAX_BINSIZ, list0, list0 + MAX_LSTSIZ - 2);
+	if (dest1 == NULL) {
+		errmsgout("NASK : BINBUF is not enough" NL);
+		return 19;
+	}
+	if (argc == 4) {
+		tmp1 = list0;
+		if (list0[MAX_LSTSIZ - 1])
+			goto over_listbuf;
+		while (*tmp1 != '\0' && tmp1 < list0 + MAX_LSTSIZ)
+			tmp1++;
+		if (GOLD_write_t(argv[3], tmp1 - list0, list0)) {
+			errmsgout("NASK : list output error" NL);
+			return 20;
+		}
+	}
+	if (argc >= 3) {
+		if (dest1 < dest0)
+			dest1 = dest0;
+		if (GOLD_write_b(argv[2], dest1 - dest0, dest0)) {
+			errmsgout("NASK : object/binary output error" NL);
+			return 21;
+		}
+	}
+
+	if (nask_errors) {
+		UCHAR strbuf[16];
+		len = nask_errors;
+		errmsgout("NASK : ");
+		src1 = &strbuf[15];
+		*src1 = '\0';
+		do {
+			*--src1 = (len % 10) + '0';
+		} while (len /= 10);
+		errmsgout(src1);
+		errmsgout(" errors." NL);
+		return 1;
+	}
+
+	return 0;
+}
+
+void GOL_sysabort(unsigned char termcode)
+{
+	static const char *termmsg[] = {
+		"",
+		"[TERM_WORKOVER]\n",
+		"[TERM_OUTOVER]\n",
+		"[TERM_ERROVER]\n",
+		"[TERM_BUGTRAP]\n",
+		"[TERM_SYSRESOVER]\n",
+		"[TERM_ABORT]\n"
+	};
+
+	GO_stderr.p1 += 128; /* äºˆå‚™ã«å–ã£ã¦ãŠã„ãŸåˆ†ã‚’å¾©æ´» */
+	/* ãƒãƒƒãƒ•ã‚¡ã‚’å‡ºåŠ› */
+	if (termcode <= 6)
+		errmsgout(termmsg[termcode]);
+	if (GO_stderr.p - GO_stderr.p0)
+		GOLD_write_t(NULL, GO_stderr.p - GO_stderr.p0, GO_stderr.p0);
+	GOLD_exit((termcode == 0) ? GOL_retcode : 1);
+}
